debug st + fix vm_set

+++ srcs/vm_exec/get_values.c
@@ -6,7 +6,7 @@
 /*   By: gwoodwar <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/10 18:40:42 by gwoodwar          #+#    #+#             */
-/*   Updated: 2016/05/13 11:49:39 by jaguillo         ###   ########.fr       */
+/*   Updated: 2016/05/17 19:54:32 by jaguillo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ uint32_t			get_value(t_vm const *vm, t_process *const process,
 	else if (arg_type == DIR_CODE)
 		return (arg);
 	else
-		return (vm_get(vm, process->reg_pc + arg % IDX_MOD, 4));
+		return (vm_get(vm, process->reg_pc + ((int32_t)arg) % IDX_MOD, 4));
 }
 
 uint32_t			get_lvalue(t_vm const *vm, t_process *const process,
